Move VM for PhotoField inside field package
