+++ gnd/src/main/java/com/google/android/gnd/ui/editobservation/field/PhotoFieldViewModel.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.android.gnd.ui.editobservation;
+package com.google.android.gnd.ui.editobservation.field;
 
 import android.net.Uri;
 import android.view.View;
